+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_params.py
@@ -394,6 +394,9 @@ def load_arguments(self, _):
     with self.argument_context('network dns zone update') as c:
         c.ignore('if_none_match')
 
+    with self.argument_context('network dns zone create') as c:
+        c.extra('parent_zone_name', options_list=['--parent-name', '-p'], help='Specify if parent zone exists for this zone and delegation for the child zone in the parent is to be added.')
+
     with self.argument_context('network dns record-set') as c:
         c.argument('target_resource', min_api='2018-05-01', help='ID of an Azure resource from which the DNS resource value is taken.')
         for item in ['record_type', 'record_set_type']: