+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/tests/latest/test_dns_commands.py
@@ -158,6 +158,36 @@ class DnsScenarioTest(ScenarioTest):
                  checks=self.is_empty())
 
     @ResourceGroupPreparer(name_prefix='cli_test_dns')
+    def test_dns_delegation(self, resource_group):
+        self.kwargs['parent_zone_name'] = 'books.com'
+        self.cmd('network dns zone create -n {parent_zone_name} -g {rg}').get_output_in_json()
+
+        base_record_sets = 2
+        self.cmd('network dns zone show -n {parent_zone_name} -g {rg}',
+                 checks=self.check('numberOfRecordSets', base_record_sets))
+
+        self.kwargs['child_zone_name'] = 'nursery.books.com'
+        child_zone = self.cmd('network dns zone create -n {child_zone_name} -g {rg} -p {parent_zone_name}').get_output_in_json()
+        child_name_server_count = len(child_zone['nameServers'])
+
+        record_sets_with_ns_delegation = 3
+        self.cmd('network dns zone show -n {parent_zone_name} -g {rg}',
+                 checks=self.check('numberOfRecordSets', record_sets_with_ns_delegation)).get_output_in_json()
+
+        record_set_name = self.kwargs['child_zone_name'].replace('.' + self.kwargs['parent_zone_name'], '')
+        self.kwargs['record_set_name'] = record_set_name
+        self.cmd('network dns record-set ns show -n {record_set_name} -g {rg} --zone-name {parent_zone_name}',
+                 checks=self.check('length(nsRecords)', child_name_server_count))
+
+        #clean up by deleting the created resources
+        self.cmd('network dns zone delete -g {rg} -n {parent_zone_name} -y',
+                 checks=self.is_empty())
+
+        self.cmd('network dns zone delete -g {rg} -n {child_zone_name} -y',
+                 checks=self.is_empty())
+
+
+    @ResourceGroupPreparer(name_prefix='cli_test_dns')
     def test_private_dns(self, resource_group):
 
         self.kwargs['zone'] = 'myprivatezone.com'