+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 * `application-gateway rewrite-rule`: Add `list-request-headers` and `list-response-headers` commands.
 * `application-gateway rewrite-rule condition`: Add `list-server-variables` command.
 * `express-route port update`: Fixed an issue where updating link state on an express-route port would throw an unknown attribute exception.
+* `dns zone create`: Add auto name server delegation in parent during child zone creation 
 
 2.3.5
 +++++